+++ src/token/TokenNftInfo.js
@@ -47,7 +47,7 @@ export default class TokenNftInfo {
      * @param {Timestamp} props.creationTime
      * @param {Uint8Array | null} props.metadata
      * @param {LedgerId|null} props.ledgerId
-     * @param {AccountId|null} props.spenderId
+     * @param {AccountId|null} props.allowanceSpenderAccountId
      */
     constructor(props) {
         /**
@@ -74,7 +74,7 @@ export default class TokenNftInfo {
 
         this.ledgerId = props.ledgerId;
 
-        this.spenderId = props.spenderId;
+        this.allowanceSpenderAccountId = props.allowanceSpenderAccountId;
 
         Object.freeze(this);
     }
@@ -102,7 +102,7 @@ export default class TokenNftInfo {
                 info.ledgerId != null
                     ? LedgerId.fromBytes(info.ledgerId)
                     : null,
-            spenderId:
+            allowanceSpenderAccountId:
                 info.spenderId != null
                     ? AccountId._fromProtobuf(info.spenderId)
                     : null,
@@ -120,7 +120,9 @@ export default class TokenNftInfo {
             metadata: this.metadata,
             ledgerId: this.ledgerId != null ? this.ledgerId.toBytes() : null,
             spenderId:
-                this.spenderId != null ? this.spenderId._toProtobuf() : null,
+                this.allowanceSpenderAccountId != null
+                    ? this.allowanceSpenderAccountId._toProtobuf()
+                    : null,
         };
     }
 
@@ -131,7 +133,7 @@ export default class TokenNftInfo {
      * @property {string} creationTime
      * @property {string | null} metadata
      * @property {string | null} ledgerId
-     * @property {string | null} spenderId
+     * @property {string | null} allowanceSpenderAccountId
      * @returns {TokenNftInfoJson}
      */
     toJson() {
@@ -141,8 +143,10 @@ export default class TokenNftInfo {
             creationTime: this.creationTime.toString(),
             metadata: this.metadata != null ? hex.encode(this.metadata) : null,
             ledgerId: this.ledgerId != null ? this.ledgerId.toString() : null,
-            spenderId:
-                this.spenderId != null ? this.spenderId.toString() : null,
+            allowanceSpenderAccountId:
+                this.allowanceSpenderAccountId != null
+                    ? this.allowanceSpenderAccountId.toString()
+                    : null,
         };
     }