+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
- * `TokenNftInfo.spenderId`
+ * `TokenNftInfo.allowanceSpenderAccountId`
 
 ## v2.17.0